Fish_n_Chips/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scipy
scikit-learn
joblib
xgboost
matplotlib
jupyter
requests
python-multipart
tqdm
seaborn
mne

# Speech Analysis Dependencies
openai>=1.55.0
webrtcvad==2.0.10
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
# numpy and scipy are already present, but ensuring versions if needed (skipping strict versioning to avoid conflicts with existing)
# numpy==1.24.3
# scipy==1.11.4

# NLP for linguistic features
spacy==3.7.2
python-Levenshtein==0.23.0

# ML (already present)
# scikit-learn==1.3.2
# joblib==1.3.2

# Database for data collection
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
